Clean `mods.cpp`
