+++ soh/soh/Enhancements/mods.h
@@ -1,5 +1,3 @@
-#include <stdint.h>
-
 #ifndef MODS_H
 #define MODS_H
 
@@ -7,16 +5,9 @@
 extern "C" {
 #endif
 
-void UpdateDirtPathFixState(int32_t sceneNum);
-void UpdateMirrorModeState(int32_t sceneNum);
-void UpdateHurtContainerModeState(bool newState);
 void PatchToTMedallions();
 void PatchCompasses();
-void UpdatePermanentHeartLossState();
-void UpdateHyperEnemiesState();
-void UpdateHyperBossesState();
-void InitMods();
-void UpdatePatchHand(); 
+void UpdatePatchHand();
 void SwitchAge();
 
 #ifdef __cplusplus
