+++ soh/soh/Enhancements/randomizer/hook_handlers.cpp
@@ -2471,3 +2471,5 @@ void RandomizerRegisterHooks() {
         }
     });
 }
+
+static RegisterShipInitFunc initFunc(RandomizerRegisterHooks);
