--- soh/soh/Enhancements/randomizer/hook_handlers.h
@@ -1,6 +1,0 @@
-#ifndef RANDOMIZER_HOOK_HANDLERS_H
-#define RANDOMIZER_HOOK_HANDLERS_H
-
-void RandomizerRegisterHooks();
-
-#endif // RANDOMIZER_HOOK_HANDLERS_H
