--- soh/soh/Enhancements/timesaver_hook_handlers.h
@@ -1,6 +1,0 @@
-#ifndef TIMESAVER_HOOK_HANDLERS_H
-#define TIMESAVER_HOOK_HANDLERS_H
-
-void TimeSaverRegisterHooks();
-
-#endif // TIMESAVER_HOOK_HANDLERS_H
