--- soh/soh/Enhancements/tts/tts.h
@@ -1,6 +1,0 @@
-#ifndef TTS_H
-#define TTS_H
-
-void RegisterTTS();
-
-#endif
